move streaming related code to streaming.go
